make sure the ports used by one instance is not used by other
Bypass.down closes the socket and this allows the OS to return the
same port on listen call. This is not desired, as the http request
might end up handled by different instance.

Create a new GenServer which keep tracks of all the open ports and
their owners and make sure the same port is not allocated to multiple
instances